+++ extension/src/core/constants/index.ts
@@ -3,5 +3,4 @@
  */
 
 export * from './logging'
-export * from './style'
 export * from './i18n'